redesign ARDB storage operation primitives (#566)
The red line through this commit is to redesign the ARDB storage primitives (copy, list, exists, delete, listIndices), such that they play well with the new cluster (in-memory model) design. By doing so several other issues can be easily resolved as well, and thus this PR resolves several issues at once. Depending on how the future slave cluster works out, we might want to refactor a few pieces here and there, but for now it looks fine as it is.

+ Fixes #549: ARDB storage primitives are now easier to use than ever. The public primitives are now: 'VdiskExists', 'CopyVdisk', 'DeleteVdisk', 'ListVdisks', 'ListBlockIndices';
+ Fixes #481: Deleting all ARDB data for a given vdisk is now as easy as a call to 'DeleteVdisk';
+ Fixes #353: When copying LBA sectors between different ARDB servers, we use 'HSCAN' to limit the sectors-per-call to '1000', making this copy primitive variant scalable;
+ Fixes #354: When copying non-deduped blocks between different ARDB servers, we use 'HSCAN' to limit the blocks-per-call to '1000', making this copy primitive variant scalable;
+ Fixes #206: make 'zeroctl copy vdisk' work for non-deduped vdisks, which use different clusters that have each a different amount of servers;
+ Fixes #559: An error gets now returned when the source- and target- vdisk have different storage type during a 'zeroctl copy vdisk' command;
+ Fixes #564: An error gets now returned when the source- and target- vdisk have the same storage type but different block sizes during a 'zeroctl copy vdisk' command;